+++ libs/components/src/locales/en-GB.ts
@@ -223,6 +223,12 @@ const enGB: Locale = {
 			selected: 'Selected',
 		},
 	},
+	pagination: {
+		previousPageLabel: 'Go to previous page',
+		nextPageLabel: 'Go to next page',
+		goToPageLabel: /* istanbul ignore next */ (index: number | string) =>
+			`Go to page ${index}`,
+	},
 };
 
 export default enGB;