+++ libs/components/CHANGELOG.md
@@ -4,7 +4,7 @@ This log was last generated on Thu, 26 Jan 2023 14:40:25 GMT and should not be m
 
 <!-- Start content -->
 
-## [4.26.0](https://github.com/Vonage/vivid-3/compare/vivid-v4.25.0...vivid-v4.26.0) (2025-06-16)
+## [4.26.0](https://github.com/Vonage/vivid-3/compare/vivid-v4.25.0...vivid-v4.26.0) (2025-06-18)
 
 
 ### Features
@@ -24,6 +24,8 @@ This log was last generated on Thu, 26 Jan 2023 14:40:25 GMT and should not be m
 * **data-grid:** fix sticky-header row property being overwritten by generate-header (VIV-2714) ([#2412](https://github.com/Vonage/vivid-3/issues/2412)) ([0368476](https://github.com/Vonage/vivid-3/commit/03684766e718ffeb79753fbbb61627dfb525e91d))
 * **dialpad:** Add localized aria-label to delete/backspace button (VIV-2693) ([#2402](https://github.com/Vonage/vivid-3/issues/2402)) ([42cb3bd](https://github.com/Vonage/vivid-3/commit/42cb3bdd373c0f0bfaa9a5d6190135548302c76b))
 * **number-field:** screen reader accessibility fixes (VIV-2694) ([#2405](https://github.com/Vonage/vivid-3/issues/2405)) ([72944ec](https://github.com/Vonage/vivid-3/commit/72944ec666cf03a0e31a5b636dacbe1dc07928cc))
+* **pagination:** accessibility fixes (VIV-2695) ([#2416](https://github.com/Vonage/vivid-3/issues/2416)) ([d63c485](https://github.com/Vonage/vivid-3/commit/d63c4855ff2c39effd1c8a3fa83436c8b706fb73))
+* **picker-field:** modify the clear button functionality to clear the selected value without closing the picker view (VIV-2696)  ([#2418](https://github.com/Vonage/vivid-3/issues/2418)) ([96025df](https://github.com/Vonage/vivid-3/commit/96025df08628ea7519b1a24bc9cfba188a0df14c))
 * **searchable-select:** fix issue with tags being cut-off on smaller width (VIV-2689) ([#2401](https://github.com/Vonage/vivid-3/issues/2401)) ([a4cbd7d](https://github.com/Vonage/vivid-3/commit/a4cbd7dcfbd0d77e8b2f49ba94c05ce602a384df))
 * **slider, range-slider:** remove focus transition (VIV-2690) ([#2397](https://github.com/Vonage/vivid-3/issues/2397)) ([e596b63](https://github.com/Vonage/vivid-3/commit/e596b638df15ffbf09fbbc642659ca05b03af83b))