+++ libs/vue-wrappers/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [4.26.0](https://github.com/Vonage/vivid-3/compare/vivid-vue-v4.25.0...vivid-vue-v4.26.0) (2025-06-16)
+## [4.26.0](https://github.com/Vonage/vivid-3/compare/vivid-vue-v4.25.0...vivid-vue-v4.26.0) (2025-06-18)
 
 
 ### Miscellaneous Chores